refactor: web client
